refactor: dead code elimination, Apply the changed architecture 2 (#60)
