+++ src/main/kotlin/app/askresume/api/extract/service/ExtractService.kt
@@ -1,13 +1,12 @@
 package app.askresume.api.extract.service
 
-import app.askresume.api.extract.dto.response.ExtractedTextResponse
+import app.askresume.api.extract.vo.ExtractedTextResponse
 import app.askresume.global.error.ErrorCode
 import app.askresume.global.error.exception.BusinessException
 import app.askresume.global.util.LoggerUtil.log
 import app.askresume.global.util.ParsingUtils
 import org.apache.pdfbox.pdmodel.PDDocument
 import org.apache.pdfbox.text.PDFTextStripper
-import org.jsoup.Jsoup
 import org.springframework.stereotype.Service
 import org.springframework.web.multipart.MultipartFile
 import java.io.IOException
@@ -19,7 +18,7 @@ class ExtractService {
     val log = log()
 
     fun pdfToText(file: MultipartFile): ExtractedTextResponse {
-        var text = ""
+        var text: String
         try {
             file.inputStream.use { inputStream ->
                 val document = PDDocument.load(inputStream)
