+++ src/main/kotlin/app/askresume/domain/generative/interview/repository/InterviewRepository.kt
@@ -1,4 +1,4 @@
-package app.askresume.domain.generative.interview
+package app.askresume.domain.generative.interview.repository
 
 import app.askresume.domain.generative.interview.model.ResultInterviewMaker
 import org.springframework.data.jpa.repository.JpaRepository
