+++ src/main/kotlin/app/askresume/domain/result/repository/ResultRepository.kt
@@ -3,5 +3,4 @@ package app.askresume.domain.result.repository;
 import app.askresume.domain.result.model.Result
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ResultRepository : JpaRepository<Result, Long> {
-}
+interface ResultRepository : JpaRepository<Result, Long>
