+++ src/main/kotlin/app/askresume/domain/result/service/ResultService.kt
@@ -29,6 +29,5 @@ class ResultService(
                         totalTokens = totalTokens,
             )
         )
-
     }
 }
